refactor, moving baklava node types to the 'nodes' folder
